Run dependabot-prs workflow on `labeled` action
By default, the `pull_request_target` event does not trigger
workflows when the event was in response to labels being
added. From
[GitHub's docs](https://docs.github.com/en/actions/reference/events-that-trigger-workflows#pull_request_target):

> By default, a workflow only runs when a `pull_request_target`'s
> activity type is `opened`, `synchronize`, or `reopened`.
> To trigger workflows for more activity types, use the `types` keyword.

Since Dependabot does not create PRs with labels but in fact
adds them separately, we want this action to run again
when labels are added so it can use the
`development-dependencies` label to automerge PRs.